# apps/api-go/internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title": "Liquor Store Security API", "description": "Go backend contract for the store owner security dashboard.", "version": "1.0.0"},
  "servers": [{"url": "http://localhost:3000"}],
...
  "paths": {
...
    "/api/v1/internal/ai/alerts": {"post":{"tags":["internal"],"summary":"Ingest an AI-generated alert and enqueue matching notifications","description":"Service-to-service endpoint enabled only when AI_INGEST_TOKEN is configured. sourceEventId is idempotent. The alert, evidence and matching delivery outbox rows commit in one transaction.","security":[{"aiIngestAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AIAlertIngestRequest"}}}},"responses":{"201":{"description":"Alert, evidence and delivery routes created"},"200":{"description":"Existing sourceEventId returned without duplicate rows"},"400":{"description":"Invalid alert or evidence contract"},"401":{"description":"Invalid AI service token"},"404":{"description":"Store, camera or zone not found"},"409":{"description":"Conflicting source event or evidence key"}}}},
    "/api/v1/auth/register": {"post":{"tags":["auth"],"summary":"Create the first owner and store","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}}},"responses":{"201":{"description":"Session created"},"409":{"description":"Email or store code exists"}}}},
    "/api/v1/auth/login": {"post":{"tags":["auth"],"summary":"Log in","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"201":{"description":"Session created"},"401":{"description":"Invalid credentials"}}}},
    "/api/v1/auth/refresh": {"post":{"tags":["auth"],"summary":"Rotate refresh session","responses":{"201":{"description":"Session refreshed"},"401":{"description":"Invalid refresh cookie"}}}},
...
    "/api/v1/auth/me": {"get":{"tags":["auth"],"summary":"Current user and stores","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Current user"},"401":{"description":"Unauthorized"}}}},
    "/api/v1/users/me": {
      "get":{"tags":["users"],"summary":"Get owner profile","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Profile"}}},
      "patch":{"tags":["users"],"summary":"Update owner profile","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["displayName"],"properties":{"displayName":{"type":"string","example":"Store Owner"}}}}}},"responses":{"200":{"description":"Updated"}}}
    },
    "/api/v1/stores": {
      "get":{"tags":["stores"],"summary":"List accessible stores","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Stores"}}},
      "post":{"tags":["stores"],"summary":"Create store","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreRequest"}}}},"responses":{"201":{"description":"Created"}}}
    },
    "/api/v1/stores/{storeId}": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"}],
      "get":{"tags":["stores"],"summary":"Get store","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Store"}}},
      "patch":{"tags":["stores"],"summary":"Update store","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"},"timezone":{"type":"string"}}}}}},"responses":{"200":{"description":"Updated"}}}
    },
    "/api/v1/stores/{storeId}/members": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"}],
      "get":{"tags":["stores"],"summary":"List store members","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Members"}}},
      "post":{"tags":["stores"],"summary":"Add existing user to store","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","role"],"properties":{"email":{"type":"string","format":"email"},"role":{"type":"string","enum":["MANAGER","OPERATOR"]}}}}}},"responses":{"201":{"description":"Added"}}}
    },
    "/api/v1/stores/{storeId}/members/{memberId}": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"name":"memberId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],
      "patch":{"tags":["stores"],"summary":"Update member role","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["role"],"properties":{"role":{"type":"string","enum":["MANAGER","OPERATOR"]}}}}}},"responses":{"200":{"description":"Updated"}}},
      "delete":{"tags":["stores"],"summary":"Remove member","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Removed"}}}
    },
    "/api/v1/stores/{storeId}/cameras": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"}],
      "get":{"tags":["cameras"],"summary":"List cameras","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Cameras"}}},
      "post":{"tags":["cameras"],"summary":"Create camera","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CameraRequest"}}}},"responses":{"201":{"description":"Created"}}}
    },
    "/api/v1/stores/{storeId}/cameras/{cameraId}": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/CameraId"}],
      "get":{"tags":["cameras"],"summary":"Get camera","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Camera"}}},
      "patch":{"tags":["cameras"],"summary":"Update camera","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CameraRequest"}}}},"responses":{"200":{"description":"Updated"}}},
      "delete":{"tags":["cameras"],"summary":"Delete camera","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Deleted"}}}
    },
    "/api/v1/stores/{storeId}/cameras/{cameraId}/zones": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/CameraId"}],
      "get":{"tags":["cameras"],"summary":"List detection zones","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Zones"}}},
      "post":{"tags":["cameras"],"summary":"Create detection zone","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ZoneRequest"}}}},"responses":{"201":{"description":"Created"}}}
    },
    "/api/v1/stores/{storeId}/cameras/{cameraId}/zones/{zoneId}": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/CameraId"},{"name":"zoneId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],
      "get":{"tags":["cameras"],"summary":"Get zone","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Zone"}}},
      "patch":{"tags":["cameras"],"summary":"Update zone","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ZoneRequest"}}}},"responses":{"200":{"description":"Updated"}}},
      "delete":{"tags":["cameras"],"summary":"Delete zone","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Deleted"}}}
    },
    "/api/v1/stores/{storeId}/alerts": {"parameters":[{"$ref":"#/components/parameters/StoreId"}],"get":{"tags":["alerts"],"summary":"List and filter alerts","security":[{"bearerAuth":[]}],"parameters":[{"name":"status","in":"query","schema":{"type":"string"}},{"name":"severity","in":"query","schema":{"type":"string"}},{"name":"type","in":"query","schema":{"type":"string"}},{"name":"cameraId","in":"query","schema":{"type":"string"}},{"name":"cursor","in":"query","schema":{"type":"string","format":"date-time"}},{"name":"limit","in":"query","schema":{"type":"integer","default":30,"maximum":100}}],"responses":{"200":{"description":"Alert page"}}}},
    "/api/v1/stores/{storeId}/alerts/{alertId}": {"parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/AlertId"}],"get":{"tags":["alerts"],"summary":"Get alert detail and evidence","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Alert detail"}}}},
    "/api/v1/stores/{storeId}/alerts/{alertId}/acknowledge": {"parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/AlertId"}],"post":{"tags":["alerts"],"summary":"Acknowledge alert","security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/AlertAction"},"responses":{"201":{"description":"Acknowledged"}}}},
    "/api/v1/stores/{storeId}/alerts/{alertId}/dismiss": {"parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/AlertId"}],"post":{"tags":["alerts"],"summary":"Dismiss alert","security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/AlertAction"},"responses":{"201":{"description":"Dismissed"}}}},
    "/api/v1/stores/{storeId}/alerts/{alertId}/resolve": {"parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/AlertId"}],"post":{"tags":["alerts"],"summary":"Resolve alert","security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/AlertAction"},"responses":{"201":{"description":"Resolved"}}}},
    "/api/v1/stores/{storeId}/alerts/{alertId}/evidence/{evidenceId}/playback-url": {"parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/AlertId"},{"$ref":"#/components/parameters/EvidenceId"}],"get":{"tags":["alerts"],"summary":"Create a short-lived evidence playback URL","description":"OPERATOR or higher. The returned bearer URL is scoped to one alert evidence object, expires shortly and never exposes the storage key.","security":[{"bearerAuth":[]}],"responses":{"201":{"description":"Short-lived playback URL"},"404":{"description":"Evidence outside alert/store scope"},"503":{"description":"Secure playback is not configured"}}}},
    "/api/v1/notification-review/{token}": {"parameters":[{"$ref":"#/components/parameters/ReviewToken"}],"get":{"tags":["alerts"],"summary":"Stream evidence through an expiring opaque review token","description":"Public bearer-token endpoint used by notification links. Supports HTTP Range and never redirects to or reveals the evidence origin/storage key.","security":[],"responses":{"200":{"description":"Evidence stream"},"206":{"description":"Partial evidence stream"},"404":{"description":"Invalid, expired or revoked token"},"502":{"description":"Evidence origin unavailable"}}}},
    "/api/v1/webhooks/whatsapp": {
      "get":{"tags":["notifications"],"summary":"Verify the Meta WhatsApp webhook","description":"Public Meta verification handshake. The configured verify token is compared server-side and is never returned.","security":[],"parameters":[{"name":"hub.mode","in":"query","required":true,"schema":{"type":"string","enum":["subscribe"]}},{"name":"hub.verify_token","in":"query","required":true,"schema":{"type":"string"}},{"name":"hub.challenge","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Challenge echoed as plain text"},"403":{"description":"Verification failed"},"503":{"description":"Webhook not configured"}}},
      "post":{"tags":["notifications"],"summary":"Receive signed WhatsApp delivery receipts","description":"Validates X-Hub-Signature-256 over the raw body, accepts at most 1 MiB, and idempotently records only stable sent/delivered/read/failed status data. Raw webhook bodies and recipient data are not persisted.","security":[],"responses":{"200":{"description":"Signed webhook accepted"},"400":{"description":"Malformed webhook payload"},"401":{"description":"Invalid signature"},"413":{"description":"Body too large"},"503":{"description":"Webhook not configured"}}}
    },
    "/api/v1/stores/{storeId}/notification-endpoints": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"}],
      "get":{"tags":["notifications"],"summary":"List notification endpoints (OWNER only)","description":"Responses never contain credentialRef (only credentialConfigured) and destinations are masked. Lists return [].","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Endpoints"},"401":{"description":"Unauthorized"},"403":{"description":"Requires store OWNER"}}},
      "post":{"tags":["notifications"],"summary":"Create notification endpoint (OWNER only)","description":"Provider is immutable afterwards. Duplicate (store, provider, destinationRef) returns 409. Raw credentials in credentialRef or config are rejected with 400. Enabling WHATSAPP additionally requires providerAccountRef (Meta Phone Number ID), an E.164 destinationRef and wabaId/templateName/templateLanguage/templateVersion/optIn evidence in config (400 otherwise).","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationEndpointCreateRequest"}}}},"responses":{"201":{"description":"Created"},"400":{"description":"Invalid payload, raw credential or WhatsApp enable validation"},"409":{"description":"Duplicate destination"}}}
    },
    "/api/v1/stores/{storeId}/notification-endpoints/{endpointId}": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/EndpointId"}],
      "patch":{"tags":["notifications"],"summary":"Update notification endpoint (OWNER only)","description":"Partial update; empty bodies and immutable fields such as provider are rejected with 400. Changing destinationRef to a duplicate returns 409. DELETE is blocked with 409 when delivery history exists - disable via isEnabled=false instead.","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationEndpointUpdateRequest"}}}},"responses":{"200":{"description":"Updated"},"400":{"description":"Invalid payload, unknown field, empty body or WhatsApp enable validation"},"404":{"description":"Endpoint outside caller's store"},"409":{"description":"Duplicate destination or delivery history"}}},
      "delete":{"tags":["notifications"],"summary":"Delete endpoint without delivery history (OWNER only)","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Deleted"},"404":{"description":"Endpoint outside caller's store"},"409":{"description":"Delivery history exists; soft-disable instead"}}}
    },
    "/api/v1/stores/{storeId}/notification-endpoints/{endpointId}/test": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/EndpointId"}],
      "post":{"tags":["notifications"],"summary":"Enqueue a TEST-kind delivery (OWNER only)","description":"Only enqueues an outbox row; no provider call happens in Task 3. Idempotent per requestId. Disabled endpoints return 409. Returns 202 Accepted with the delivery id and status only.","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationTestRequest"}}}},"responses":{"202":{"description":"Test delivery enqueued"},"400":{"description":"Missing or invalid requestId"},"404":{"description":"Endpoint outside caller's store"},"409":{"description":"Endpoint disabled"}}}
    },
    "/api/v1/stores/{storeId}/notification-rules": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"}],
      "get":{"tags":["notifications"],"summary":"List notification rules (OWNER only)","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Rules"}}},
      "post":{"tags":["notifications"],"summary":"Create notification rule (OWNER only)","description":"name is required. Omitted fields default: minimumSeverity=CRITICAL, alertTypes=[], cooldownSeconds=0, isEnabled=true. minimumSeverity and alertTypes combine with AND; empty alertTypes matches every alert type passing severity. Nonzero cooldownSeconds is evaluated transactionally against the exact incident fingerprint; emergency rules should remain 0.","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationRuleCreateRequest"}}}},"responses":{"201":{"description":"Created"},"400":{"description":"Invalid payload"}}}
    },
    "/api/v1/stores/{storeId}/notification-rules/{ruleId}": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/RuleId"}],
      "patch":{"tags":["notifications"],"summary":"Update notification rule (OWNER only)","description":"Partial update; omitted fields keep their current values and create defaults are not reapplied. Empty bodies are rejected.","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationRuleUpdateRequest"}}}},"responses":{"200":{"description":"Updated"},"400":{"description":"Invalid payload, unknown field or empty body"},"404":{"description":"Rule outside caller's store"}}},
      "delete":{"tags":["notifications"],"summary":"Delete rule without delivery history (OWNER only)","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Deleted"},"404":{"description":"Rule outside caller's store"},"409":{"description":"Delivery history exists; soft-disable instead"}}}
    },
    "/api/v1/stores/{storeId}/notification-rules/{ruleId}/channels": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/RuleId"}],
      "get":{"tags":["notifications"],"summary":"List rule channels ordered by priority (OWNER only)","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Channels"}}},
      "post":{"tags":["notifications"],"summary":"Add a channel route to a rule (OWNER only)","description":"Endpoint must belong to the same store. Lower priority routes first. Duplicate rule+endpoint or duplicate priority returns 409. Task 3 configures routing only; fallback activation after failures belongs to Task 5.","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationRuleChannelCreateRequest"}}}},"responses":{"201":{"description":"Created"},"400":{"description":"Invalid payload"},"404":{"description":"Rule or endpoint outside caller's store"},"409":{"description":"Duplicate endpoint or priority in rule"}}}
    },
    "/api/v1/stores/{storeId}/notification-rules/{ruleId}/channels/{channelId}": {
      "parameters":[{"$ref":"#/components/parameters/StoreId"},{"$ref":"#/components/parameters/RuleId"},{"$ref":"#/components/parameters/ChannelId"}],
      "patch":{"tags":["notifications"],"summary":"Update channel priority, delay or enabled flag (OWNER only)","description":"Partial update; endpointId is immutable and unknown fields are rejected with 400. Duplicate priority returns 409.","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationRuleChannelUpdateRequest"}}}},"responses":{"200":{"description":"Updated"},"400":{"description":"Invalid payload, unknown field or empty body"},"404":{"description":"Channel outside caller's store/rule"},"409":{"description":"Duplicate priority in rule or delivery history"}}},
      "delete":{"tags":["notifications"],"summary":"Delete channel without delivery history (OWNER only)","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Deleted"},"404":{"description":"Channel outside caller's store/rule"},"409":{"description":"Delivery history exists; soft-disable instead"}}}
    },
...
  },
  "components": {
    "securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"},"aiIngestAuth":{"type":"http","scheme":"bearer","bearerFormat":"opaque service token"}},
    "parameters":{
      "StoreId":{"name":"storeId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
      "CameraId":{"name":"cameraId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
      "AlertId":{"name":"alertId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
      "EndpointId":{"name":"endpointId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
      "RuleId":{"name":"ruleId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
...
      ,"EvidenceId":{"name":"evidenceId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}
...
      ,"ReviewToken":{"name":"token","in":"path","required":true,"schema":{"type":"string","minLength":40,"maxLength":64}}
    },
    "requestBodies":{"AlertAction":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"note":{"type":"string","maxLength":1000,"example":"Reviewed by owner"}}}}}}},
    "schemas":{
      "LoginRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email","example":"owner@demo.com"},"password":{"type":"string","format":"password"}}},
      "RegisterRequest":{"type":"object","required":["email","password","displayName","storeName","storeCode"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password","minLength":12},"displayName":{"type":"string"},"storeName":{"type":"string"},"storeCode":{"type":"string","pattern":"^[a-z0-9-]{3,40}$"},"address":{"type":"string"}}},
      "StoreRequest":{"type":"object","required":["name","code"],"properties":{"name":{"type":"string"},"code":{"type":"string"},"address":{"type":"string"}}},
      "CameraRequest":{"type":"object","required":["name","location","protocol","streamGatewayRef"],"properties":{"name":{"type":"string"},"location":{"type":"string"},"protocol":{"type":"string","enum":["RTSP","ONVIF","HLS","WEBRTC"]},"streamGatewayRef":{"type":"string","description":"Safe gateway reference, never an RTSP URL or password"},"status":{"type":"string","enum":["ONLINE","OFFLINE","DISABLED"]},"isEnabled":{"type":"boolean"}}},
      "ZoneRequest":{"type":"object","required":["name","kind","polygon"],"properties":{"name":{"type":"string"},"kind":{"type":"string","enum":["HIGH_VALUE","CASHIER","STOCKROOM","ENTRANCE","CUSTOM"]},"expectedPersonCategory":{"type":"string","enum":["EMPLOYEE","CUSTOMER","UNKNOWN"]},"polygon":{"type":"array","minItems":3,"items":{"type":"array","minItems":2,"maxItems":2,"items":{"type":"number","minimum":0,"maximum":1}},"example":[[0.1,0.1],[0.9,0.1],[0.9,0.9],[0.1,0.9]]},"dwellThresholdSeconds":{"type":"integer","minimum":1}}},
      "AIAlertIngestRequest":{"type":"object","required":["sourceEventId","storeId","cameraId","type","severity","detectedAt","evidence"],"properties":{"sourceEventId":{"type":"string","minLength":1,"maxLength":200},"correlationId":{"type":"string","maxLength":200},"storeId":{"type":"string","format":"uuid"},"cameraId":{"type":"string","format":"uuid"},"zoneId":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["CASH_DRAWER_WITHOUT_CUSTOMER","SUSPICIOUS_CASH_HANDLING","POS_VOID_OR_REFUND","UNAUTHORIZED_STOCKROOM_ACCESS","HIGH_VALUE_ZONE_DWELL","WEAPON_DETECTED","SUSPICIOUS_PRODUCT_CONCEALMENT"]},"severity":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]},"subjectPersonCategory":{"type":"string","enum":["EMPLOYEE","CUSTOMER","UNKNOWN"],"default":"UNKNOWN"},"confidence":{"type":"number","minimum":0,"maximum":1},"detectedAt":{"type":"string","format":"date-time"},"observedStartAt":{"type":"string","format":"date-time"},"observedEndAt":{"type":"string","format":"date-time"},"metadata":{"type":"object","additionalProperties":true,"maxProperties":100},"evidence":{"type":"array","minItems":1,"maxItems":5,"items":{"$ref":"#/components/schemas/AIAlertEvidence"}}}},
...
      "NotificationEndpointUpdateRequest":{"type":"object","description":"Partial update; provider is immutable and unknown or empty bodies are rejected with 400.","properties":{"label":{"type":"string","minLength":1,"maxLength":120},"providerAccountRef":{"type":"string"},"destinationRef":{"type":"string"},"credentialRef":{"type":"string","description":"Secret-manager reference; raw tokens are rejected"},"config":{"type":"object","additionalProperties":true},"isEnabled":{"type":"boolean"}}},
...
    }
  }
}